Fixed encodier not supporting wrap around diffs

+++ Src/QOI.Core/Chunk/DiffReader.cs
@@ -27,7 +27,7 @@ internal class DiffReader : IChunkReader
         int rDiff = TruncateToBits((chunk >> 4), 2) + MinDiff;
         int gDiff = TruncateToBits((chunk >> 2), 2) + MinDiff;
         int bDiff = TruncateToBits(chunk, 2) + MinDiff;
-        return new(0, (short)rDiff, (short)gDiff, (short)bDiff);
+        return new(0, (sbyte)rDiff, (sbyte)gDiff, (sbyte)bDiff);
     }
 
     private static int TruncateToBits(int number, byte bitCount)
