+++ Src/QOI.Core/Chunk/LumaReader.cs
@@ -18,11 +18,11 @@ internal class LumaReader : IChunkReader
     private QoiColorDiff ReadDiff(ReadOnlySpan<byte> chunk)
     {
         // 6-bit green channel difference from the previous pixel -32..31
-        short gDiff = (short)(Tag.LUMA.Erase(chunk[0]) - 32);
+        sbyte gDiff = (sbyte)(Tag.LUMA.Erase(chunk[0]) - 32);
         // 4-bit red channel difference minus green channel difference -8..7
-        short rDiff = (short)(gDiff + (chunk[1] >> 4) - 8);
+        sbyte rDiff = (sbyte)(gDiff + (chunk[1] >> 4) - 8);
         // 4-bit blue channel difference minus green channel difference -8..7
-        short bDiff = (short)(gDiff + (chunk[1] & 0b1111) - 8);
+        sbyte bDiff = (sbyte)(gDiff + (chunk[1] & 0b1111) - 8);
         return new QoiColorDiff(0, rDiff, gDiff, bDiff);
     }
 }
